Refactoring: move router
